pyproject.toml
[tool.poetry]
name = "table_builder"
version = "0.1.0"
description = ""
authors = ["kvarel <a.parshin@comearth.ru>"]

[tool.poetry.dependencies]
python = "3.9"
django = "3.2"
django-environ = "^0.9.0"
whitenoise = "^6.3.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
SQLAlchemy = "^1.4.46"
cryptography = "^39.0.0"
djangorestframework-csv = "^2.1.1"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.1"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
